Parallel isodoc executions: https://github.com/metanorma/metanorma/issues/220
